split review reaction layer
