+++ apps/ios/Flashcards/Flashcards/Review/ReviewReactionColors.swift
@@ -1,0 +1,31 @@
+import SwiftUI
+
+extension ReviewReactionRenderer {
+    static func reviewReactionRedColor() -> Color {
+        Color(red: 0.96, green: 0.13, blue: 0.17)
+    }
+
+    static func reviewReactionOrangeColor() -> Color {
+        Color(red: 1.00, green: 0.50, blue: 0.10)
+    }
+
+    static func reviewReactionYellowColor() -> Color {
+        Color(red: 1.00, green: 0.82, blue: 0.14)
+    }
+
+    static func reviewReactionGreenColor() -> Color {
+        Color(red: 0.12, green: 0.76, blue: 0.34)
+    }
+
+    static func reviewReactionBlueColor() -> Color {
+        Color(red: 0.16, green: 0.62, blue: 1.00)
+    }
+
+    static func reviewReactionPurpleColor() -> Color {
+        Color(red: 0.58, green: 0.34, blue: 0.98)
+    }
+
+    static func reviewReactionPinkColor() -> Color {
+        Color(red: 1.00, green: 0.32, blue: 0.68)
+    }
+}
